huggingface/requirements.txt
fasthtml-hf
python-fasthtml